+++ smif/runner.py
@@ -52,68 +52,3 @@ class Asset(object):
             The attributes of the asset
         """
         return self._attributes
-
-
-class ModelRunner(object):
-    """Contains the data and methods associated with the smif facing aspects of
-    running a sector model
-
-    ModelRunner expects to find a yaml configuration file containing
-    - lists of assets in ``models/<model_name>/asset_*.yaml``
-    - structure of attributes in ``models/<model_name/<asset_name>.yaml
-
-    ModelRunner expects to find a ``run.py`` file in ``models/<model_name>``.
-    ``run.py`` contains a python script which subclasses
-    :class:`smif.sectormodel.SectorModel` to wrap the sector model.
-
-    """
-    def __init__(self, model_name, attributes):
-        """
-        Arguments
-        =========
-        project_folder: str
-            The path to the project folder
-        model_name : str
-            Name of the model
-        """
-        self._model_name = model_name
-        self._attributes = attributes
-
-    @property
-    def name(self):
-        """The name of the sector model
-
-        Returns
-        =======
-        str
-            The name of the sector model
-
-        Note
-        ====
-        The name corresponds to the name of the folder in which the
-        configuration is expected to be found
-
-        """
-        return self._model_name
-
-    @property
-    def assets(self):
-        """The names of the assets
-
-        Returns
-        =======
-        list
-            A list of the names of the assets
-        """
-        return sorted([asset for asset in self._attributes.keys()])
-
-    @property
-    def attributes(self):
-        """The collection of asset attributes
-
-        Returns
-        =======
-        dict
-            The collection of asset attributes
-        """
-        return self._attributes
